+++ smarts/core/agent_interface.py
@@ -320,11 +320,14 @@ class AgentInterface:
     The choice of action space; this also decides the controller that will be enabled and the chassis type that will be used.
     """
 
-    vehicle_type: Literal[
-        "bus", "coach", "motorcycle", "pedestrian", "sedan", "trailer", "truck"
-    ] = "sedan"
+    vehicle_type: str = ""
     """
     The choice of vehicle type.
+    """
+
+    vehicle_class: str = "generic_sedan"
+    """
+    The choice of vehicle classes from the vehicle definition list.
     """
 
     accelerometer: Union[Accelerometer, bool] = True
@@ -369,15 +372,17 @@ class AgentInterface:
             self.lane_positions, LanePositions
         )
         self.signals = AgentInterface._resolve_config(self.signals, Signals)
-        assert self.vehicle_type in {
-            "bus",
-            "coach",
-            "motorcycle",
-            "pedestrian",
-            "sedan",
-            "trailer",
-            "truck",
-        }
+        if self.vehicle_type != "":
+            warnings.warn(
+                "`vehicle_type` is now deprecated. Instead use `vehicle_class`.",
+                category=DeprecationWarning,
+            )
+            assert self.vehicle_type in (
+                "sedan",
+                "bus",
+            ), "Only these values were supported at deprecation."
+            self.vehicle_class = self.vehicle_type
+        assert isinstance(self.vehicle_class, str)
 
     @staticmethod
     def from_type(requested_type: AgentType, **kwargs):